pyproject.toml
[tool.poetry]
name = "django-example"
version = "0.1.0"
description = ""
authors = ["a4rcvv <35623957+a4rcvv@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "django_example"}]

[tool.poetry.dependencies]
python = "3.10.*"
django = "4.1.9"
django-configurations = "^2.4.1"
django-filter = "^23.2"
gunicorn = "^20.1.0"
django-environ = "^0.10.0"
werkzeug = "^2.3.4"
drf-spectacular = "^0.26.2"
psycopg2 = "^2.9.6"
django-extensions = "^3.2.1"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.1.0"
rules = "^3.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
django-debug-toolbar = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
DJANGO_CONFIGURATION = "Local"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.isort]
profile = "black"

[tool.black]
target-version = ["py311"]
